remove elm architecture examples
so Elm's solution was indeed to use an int, but as an id, not as an
index. I'm not convinced this is much better, but let's roll with it.

I think reading about the Elm architecture was still quite useful. Dare
I say it? It encapsulates state and methods, and yet is immutable: this
is OOP done right!